requirements.txt
amqp==2.5.1
awscli==1.16.238
backports.csv==1.0.7
billiard==3.6.1.0
boto3==1.9.225
botocore==1.12.228
celery==4.3.0
certifi==2019.9.11
colorama==0.3.9
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.5
django-appconf==1.0.3
django-image-cropping==1.2.0
django-import-export==1.2.0
django-storages==1.7.2
docutils==0.15.2
easy-thumbnails==2.6
et-xmlfile==1.0.1
gunicorn==19.9.0
importlib-metadata==0.20
jdcal==1.4.1
jmespath==0.9.4
kombu==4.6.4
meld3==2.0.0
more-itertools==7.2.0
nltk==3.4.5
odfpy==1.4.0
openpyxl==2.6.3
Pillow==6.1.0
psycopg2-binary==2.8.3
pyasn1==0.4.7
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
PyYAML==5.1.2
rake-nltk==1.0.4
rsa==3.4.2
s3transfer==0.2.1
sentry-sdk==0.11.2
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
uWSGI==2.0.18
vine==1.3.0
whitenoise==4.1.4
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
